Defensive programming has been implemented in the command for series and movies.
Until now, only the length of the message to be sent was controlled in the subtitle command. In case there were too many options in the series or movies when displaying the results, this could cause an internal error and no response would be sent. The message is now split into several messages to avoid this problem and the bot has no errors in that scenario.