Changed the logic in the command to search for series when prioritising the results.
Now, the first results will always be full season packs. The following results will correspond to the requested episode. This is done to avoid showing only packs that may not contain subtitles, so if you want to get the individual episode with subtitles it is now possible.

Updated decorator dependency from 5.0.6 to 5.0.7.